backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pypdf>=4.0.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
openai>=1.10.0
langchain>=0.1.0
langchain-community>=0.4.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
httpx>=0.27.0
# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

